TYCHO-380 can't run tests against eclipse 3.6 target platform
Due to a bug in maven-osgi-test-plugin, test OSGi runtime included only one
version of bundles with the same symbolic name. The exact version used was
generally random and depended on sort order of LinkedHashMap.

In eclipse 3.6, there are two versions of org.junit bundle, 3.8.x and 4.x.x,
but only one semi-random version was used at runtime due to the bug described
above. As result, some test projects executed successfully, but some failed
because test runner picked to execute tests did not match org.junit version
available in test runtime.

The fix was to allow multiple versions of bundles with the same symbolic
name in the test OSGi runtime. This unfolded into a major refactoring
of TargetPlatform interface, which became very generic collection of
ArtifactDescriptors with very simple query API.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>