TYCHO-9 set compiler src/target level according to MANIFEST
set compiler source and target level according to
Bundle-RequiredExecutionEnvironment in MANIFEST.MF.
Includes unit tests and integration tests for the compiler.
In case several execution environments are set, minimum
source and target version wins.
POM/MANIFEST precedence for setting source and target
level is MANIFEST wins (with an override warning if
source/target was set to a different value in POM).