Fix: qualifier version is logged for POM files instead of SNAPSHOT version
Inside the log we find a statement that says: x.y.z-qualifier has been updated with a.b.c-SNAPSHOT for POM files. Of course we do not have qualifier versions inside POMs.
The version is remembered inside the Change instances in a canonical representation based on qualifiers. This needs to be translated to POM before the log message is written.