TYCHO-444 non-java source files should be configurable
for non-java source files like AspectJ, OSGi compiler
mojo can now be configured to not copy these files
to the output folder because they are treated as resources.
Use the excludeResources configuration parameter
of maven-osgi-compiler-plugin to define exclusion filters
for non-java resource files which should not
be copied to the output directory.
Includes unit test.