TYCHO-473 Bind OSGiSourceMojo to prepare-package phase by default
Instead of "package", bind OsgiSourceMojo to "prepare-package" phase
by default. This is needed because P2MetadataMojo is also bound to
phase "package" by components.xml and depends on the result of
OSGiSourceMojo for the case that p2 metadata has to be
generated for a source bundle jar.
Existing integration tests for TYCHO-192 and TYCHO-449
are adapted to reproduce the error in case the fix in
OSGiSourceMojo is not applied.
Actually OSGiSourceMojo only worked if explicitly bound to
a pre-package phase in pom.xml before this fix.