TYCHO-432 enable configuration of failIfNoTests for maven-osgi-test-plugin
maven-osgi-test-plugin can now be configured to not fail
when there are no tests to run.
This can be done by either adding -DfailIfNoTests=false as a argument
to the command line or by adding a <failIfNoTests>false</failIfNoTests>
element into the configuration section of maven-osgi-test-plugin.
In contrary to maven-surefire-plugin, the default value is true.
This is because
- it's backwards compatible with previous behaviour
- tycho has separate dedicated test modules which
  should contain tests if they exist.