TYCHO-188 Demo p2 updateable RCP application
Add a reactor in tycho-demo/itp04-rcp which shows how
to build a p2-updateable RCP application using
the packaging type eclipse-repository.
The demo product is a hello world RCP which
can be updated via "Help > Check for updates".
The demo builds ready-to-use distribution zips for
win32 and linux. Details see README.txt.