Fail build if <environment> and <environments> tags are mixed
The target-platform-configuration still supports an <environment>
element as direct child of the <configuration> element - although
with a deprecation warning. If a parent or child POM uses the
recommended <configurations> (with 's') element, this can lead to
very obscure effects, because the two configurations do not
override each other. To prevent this obscure case, mixing
<environment> and <environments> configuration is prohibited.