TYCHO-483 add maven dependencies for checked-in jars
Dependencies to nested jars within the same reactor
are now added as system scope dependencies to
maven dependency model if they exist in the filesystem.
This covers the scenario of jars checked in to SCM.
Move corresponding testcase from maven-osgi-compiler-plugin
where it was misplaced to a dedicated testcase in
tycho-osgi-components.