TYCHO-95 generate rudimentary mojo docs
add a build profile for invoking mvn site.
Generated site includes has only mojo docs
for end-user relevant maven plugins for now.
Modules tycho-maven-plugin and maven-tycho-plugin
cuase mvn site to fail so move them to a
separate build profile which is deactivated
when the site profile is activated.
Generated docs need to be reviewed and samples
should be added before we can publish it, so
this is just a starting point.