340852 build fails when project path contains symlinks
On *nix, the build would fail with NPE in EquinoxResolver
if mvn is started with a file argument containing a symbolic
link.
This was because BundleDescriptions were created using the
canonical path of the project basedir but the lookup was
using the non-canonicalized path.
Consistently use canonicalized path for BundleDescription lookup.